#1262  often hear the sound of breaking bone (http://www.soundsnap.com/foley_bone_break_with_fast_twist_2 ) _ I don’t know why. __
fixed avoiding, hunting and escaping layers. fish movements now more realistic 
fixed bug with trembling boost sound in editor
fixed bug with trembling man hurt sound

+++ Assets/_scripts/menus/HighScores.cs
@@ -109,8 +109,8 @@ public class HighScores : MonoBehaviour {
 			GUI.Label(new Rect(rect.x + (rect.width - 90), rect.y, 90, rect.height), string.Format("{0,1:00.0}", float.Parse(weight)), textStyle);
 		}
 		
-		public int getWeight() {
-			return int.Parse(weight);
+		public float getWeight() {
+			return float.Parse(weight);
 		}
 		
 		public void Copy(PlayerScore arg) {
